requirements.txt
flask
flask_jwt_extended
flask_socketio
flask_pymongo